refactor(tests): move test helper types into TestTypes folder
