+++ test/Atc.Rest.Client.Tests/EndpointResponseTests.cs
@@ -419,23 +419,4 @@ public class EndpointResponseTests
         // Assert
         sut.Content.Should().Be(expectedContent);
     }
-
-    private sealed class TestableEndpointResponse : EndpointResponse
-    {
-        public TestableEndpointResponse(
-            bool isSuccess,
-            HttpStatusCode statusCode,
-            string content,
-            object? contentObject,
-            IReadOnlyDictionary<string, IEnumerable<string>> headers)
-            : base(isSuccess, statusCode, content, contentObject, headers)
-        {
-        }
-
-        public InvalidOperationException GetInvalidContentAccessException<TExpected>(
-            HttpStatusCode expectedStatusCode,
-            string propertyName)
-            where TExpected : class
-            => InvalidContentAccessException<TExpected>(expectedStatusCode, propertyName);
-    }
 }
