+++ test/Atc.Rest.Client.Tests/StreamBinaryEndpointResponseTests.cs
@@ -206,23 +206,4 @@ public sealed class StreamBinaryEndpointResponseTests
         sut.IsSuccess.Should().BeTrue();
         sut.ErrorContent.Should().BeNull();
     }
-
-    private sealed class TestableStreamBinaryEndpointResponse : StreamBinaryEndpointResponse
-    {
-        public TestableStreamBinaryEndpointResponse(
-            bool isSuccess,
-            HttpStatusCode statusCode,
-            Stream? contentStream,
-            string? contentType,
-            string? fileName,
-            long? contentLength)
-            : base(isSuccess, statusCode, contentStream, contentType, fileName, contentLength)
-        {
-        }
-
-        public InvalidOperationException GetInvalidContentAccessException(
-            HttpStatusCode expectedStatusCode,
-            string propertyName)
-            => InvalidContentAccessException(expectedStatusCode, propertyName);
-    }
 }
