+++ test/Atc.Rest.Client.Tests/StreamingEndpointResponseTests.cs
@@ -161,22 +161,4 @@ public sealed class StreamingEndpointResponseTests
         // Assert
         exception.Message.Should().Contain("OKContent");
     }
-
-    private sealed class TestableStreamingEndpointResponse<T> : StreamingEndpointResponse<T>
-    {
-        public TestableStreamingEndpointResponse(
-            bool isSuccess,
-            HttpStatusCode statusCode,
-            IAsyncEnumerable<T?>? content,
-            string? errorContent,
-            HttpResponseMessage? httpResponse)
-            : base(isSuccess, statusCode, content, errorContent, httpResponse)
-        {
-        }
-
-        public InvalidOperationException GetInvalidContentAccessException(
-            HttpStatusCode expectedStatusCode,
-            string propertyName)
-            => InvalidContentAccessException(expectedStatusCode, propertyName);
-    }
 }
