+++ test/Atc.Rest.Client.Tests/TestTypes/CircularModel.cs
@@ -1,0 +1,6 @@
+namespace Atc.Rest.Client.Tests.TestTypes;
+
+public sealed class CircularModel
+{
+    public CircularModel? Self { get; set; }
+}
