+++ test/Atc.Rest.Client.Tests/TestTypes/DateTimeModel.cs
@@ -1,0 +1,3 @@
+namespace Atc.Rest.Client.Tests.TestTypes;
+
+public sealed record DateTimeModel(DateTimeOffset Timestamp);
