+++ test/Atc.Rest.Client.Tests/TestTypes/EmptyModel.cs
@@ -1,0 +1,6 @@
+namespace Atc.Rest.Client.Tests.TestTypes;
+
+[SuppressMessage("Major Code Smell", "S2094:Classes should not be empty", Justification = "Test helper type")]
+public sealed class EmptyModel
+{
+}
