+++ test/Atc.Rest.Client.Tests/TestTypes/NestedModel.cs
@@ -1,0 +1,3 @@
+namespace Atc.Rest.Client.Tests.TestTypes;
+
+public sealed record NestedModel(TestModel? Parent, TestModel? Child);
