+++ test/Atc.Rest.Client.Tests/TestTypes/StatusContainer.cs
@@ -1,0 +1,3 @@
+namespace Atc.Rest.Client.Tests.TestTypes;
+
+public sealed record StatusContainer(TestStatus Status);
