+++ test/Atc.Rest.Client.Tests/TestTypes/TestModel.cs
@@ -1,0 +1,3 @@
+namespace Atc.Rest.Client.Tests.TestTypes;
+
+public sealed record TestModel(string Name, int Value);
