+++ test/Atc.Rest.Client.Tests/TestTypes/TestOptions.cs
@@ -1,0 +1,6 @@
+namespace Atc.Rest.Client.Tests.TestTypes;
+
+[SuppressMessage("Major Code Smell", "S2094:Classes should not be empty", Justification = "Test helper type")]
+internal sealed class TestOptions : AtcRestClientOptions
+{
+}
