+++ test/Atc.Rest.Client.Tests/TestTypes/TestStatus.cs
@@ -1,0 +1,8 @@
+namespace Atc.Rest.Client.Tests.TestTypes;
+
+public enum TestStatus
+{
+    Inactive,
+    Active,
+    Pending,
+}
